Refactor tests, update jbx
