+++ requirements.txt
@@ -1,3 +1,3 @@
 requests
-jbxapi==2.10.1
+jbxapi
 xmltodict
