+++ tests/resources/joe_submission_new.json
@@ -1,0 +1,5 @@
+{
+    "data": {
+        "submission_id": "100001"
+    }
+}
