+++ tests/test_wildfire.py
@@ -1,0 +1,10 @@
+import io
+import os
+import json
+import unittest
+try:
+    from unittest.mock import patch, ANY as MOCK_ANY
+except ImportError:
+    from mock import patch, ANY as MOCK_ANY
+import responses
+from . import read_resource
